refactor: simplify compression implementation
- Remove LZ-string backward compatibility
- Remove Web Worker implementation
- Simplify code to use only gzip compression
- Clean up unused dependencies and files
