--- app/utils/compression-advanced.ts
@@ -1,251 +1,0 @@
-import pako from 'pako'
-import LZString from 'lz-string'
-
-// Check for native Compression Streams API support
-export const hasNativeCompression = (() => {
-  if (typeof window === 'undefined') return false
-  try {
-    new CompressionStream('gzip')
-    new DecompressionStream('gzip')
-    return true
-  } catch {
-    return false
-  }
-})()
-
-// Performance thresholds
-export const COMPRESSION_THRESHOLDS = {
-  SYNC_MAX_SIZE: 50 * 1024, // 50KB - use sync compression below this
-  LARGE_DOC_SIZE: 100 * 1024, // 100KB - show warning above this
-  MAX_URL_LENGTH: 32000, // Safe for modern browsers (was 2000)
-  INSTANT_THRESHOLD: 100, // 100ms - perceived as instant
-} as const
-
-// Convert string to Uint8Array
-function stringToUint8Array(str: string): Uint8Array {
-  return new TextEncoder().encode(str)
-}
-
-
-// Base64 URL-safe encoding
-export function toBase64URL(uint8Array: Uint8Array): string {
-  // Convert to regular base64
-  let binary = ''
-  uint8Array.forEach(byte => binary += String.fromCharCode(byte))
-  let base64 = btoa(binary)
-  
-  // Make URL-safe
-  return base64
-    .replace(/\+/g, '-')
-    .replace(/\//g, '_')
-    .replace(/=+$/, '') // Remove padding
-}
-
-// Base64 URL-safe decoding
-export function fromBase64URL(str: string): Uint8Array {
-  // Add padding if needed
-  while (str.length % 4) str += '='
-  
-  // Convert from URL-safe to regular base64
-  str = str.replace(/-/g, '+').replace(/_/g, '/')
-  
-  // Decode
-  const binary = atob(str)
-  const len = binary.length
-  const bytes = new Uint8Array(len)
-  for (let i = 0; i < len; i++) {
-    bytes[i] = binary.charCodeAt(i)
-  }
-  return bytes
-}
-
-// Native compression using Compression Streams API
-async function compressNative(text: string): Promise<Uint8Array> {
-  const cs = new CompressionStream('gzip')
-  const byteData = stringToUint8Array(text)
-  const stream = new Response(byteData).body!
-  const compressedStream = stream.pipeThrough(cs)
-  const compressedBlob = await new Response(compressedStream).blob()
-  return new Uint8Array(await compressedBlob.arrayBuffer())
-}
-
-// Native decompression using Decompression Streams API
-async function decompressNative(data: Uint8Array): Promise<string> {
-  const ds = new DecompressionStream('gzip')
-  const stream = new Response(data).body!
-  const decompressedStream = stream.pipeThrough(ds)
-  const decompressedBlob = await new Response(decompressedStream).blob()
-  return await decompressedBlob.text()
-}
-
-// Pako compression fallback
-function compressPako(text: string): Uint8Array {
-  return pako.gzip(text, { level: 6 })
-}
-
-// Pako decompression fallback
-function decompressPako(data: Uint8Array): string {
-  return pako.ungzip(data, { to: 'string' })
-}
-
-// Main compression function with automatic method selection
-export async function compressContent(
-  content: string,
-  options: { async?: boolean } = {}
-): Promise<string> {
-  const startTime = performance.now()
-  
-  try {
-    let compressed: Uint8Array
-    
-    // Decide whether to use async based on size
-    const shouldUseAsync = options.async !== false && 
-                          content.length > COMPRESSION_THRESHOLDS.SYNC_MAX_SIZE
-    
-    if (hasNativeCompression && shouldUseAsync) {
-      // Use native async compression for large documents
-      compressed = await compressNative(content)
-    } else if (hasNativeCompression && !shouldUseAsync) {
-      // For small documents, we can still use native but synchronously
-      // by using a smaller chunk size
-      compressed = await compressNative(content)
-    } else {
-      // Fallback to Pako (synchronous)
-      compressed = compressPako(content)
-    }
-    
-    const encoded = toBase64URL(compressed)
-    
-    // Log performance metrics
-    const duration = performance.now() - startTime
-    console.debug('Compression stats:', {
-      originalSize: content.length,
-      compressedSize: compressed.length,
-      encodedSize: encoded.length,
-      ratio: Math.round((1 - compressed.length / content.length) * 100) + '%',
-      duration: Math.round(duration) + 'ms',
-      method: hasNativeCompression ? 'native' : 'pako'
-    })
-    
-    return encoded
-  } catch (error) {
-    console.error('Compression failed:', error)
-    throw new Error('Failed to compress content')
-  }
-}
-
-// Main decompression function with format detection
-export async function decompressContent(compressed: string): Promise<string | null> {
-  try {
-    // First, check if it's LZ-string format (for backward compatibility)
-    if (compressed.includes('_') && !compressed.includes('-')) {
-      // Likely LZ-string format (uses _ but not -)
-      return LZString.decompressFromEncodedURIComponent(compressed)
-    }
-    
-    // Decode from base64 URL
-    const data = fromBase64URL(compressed)
-    
-    // Check for gzip magic bytes (1f 8b)
-    if (data[0] === 0x1f && data[1] === 0x8b) {
-      // It's gzip compressed
-      if (hasNativeCompression) {
-        return await decompressNative(data)
-      } else {
-        return decompressPako(data)
-      }
-    }
-    
-    // Fallback: try LZ-string in case detection failed
-    return LZString.decompressFromEncodedURIComponent(compressed)
-  } catch (error) {
-    console.error('Decompression failed:', error)
-    // Last resort: try LZ-string directly
-    try {
-      return LZString.decompressFromEncodedURIComponent(compressed)
-    } catch {
-      return null
-    }
-  }
-}
-
-// Estimate compressed size without actually compressing
-export function estimateCompressedSize(content: string): number {
-  // Rough estimation based on typical markdown compression ratios
-  // Gzip typically achieves 70-80% compression on markdown
-  const estimatedCompressed = content.length * 0.25
-  // Base64 encoding adds ~33% overhead
-  const estimatedEncoded = estimatedCompressed * 1.33
-  return Math.round(estimatedEncoded)
-}
-
-// Check if content can be shared via URL
-export function canShareViaURL(content: string): boolean {
-  const baseURL = window.location.origin + '/#share/'
-  const estimatedSize = estimateCompressedSize(content)
-  return (baseURL.length + estimatedSize) < COMPRESSION_THRESHOLDS.MAX_URL_LENGTH
-}
-
-// Progress callback type for worker compression
-export type CompressionProgress = {
-  status: 'compressing' | 'encoding' | 'done' | 'error'
-  progress: number // 0-100
-  message?: string
-}
-
-// Compress in Web Worker (for very large documents)
-export async function compressInWorker(
-  content: string,
-  onProgress?: (progress: CompressionProgress) => void
-): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker('/compression-worker.js')
-    const id = Math.random().toString(36).substring(7)
-    
-    worker.addEventListener('message', (event) => {
-      if (event.data.id !== id) return
-      
-      switch (event.data.type) {
-        case 'progress':
-          onProgress?.({
-            status: event.data.status,
-            progress: event.data.progress
-          })
-          break
-          
-        case 'result':
-          worker.terminate()
-          console.debug('Worker compression stats:', event.data.data.stats)
-          resolve(event.data.data.encoded)
-          break
-          
-        case 'error':
-          worker.terminate()
-          onProgress?.({
-            status: 'error',
-            progress: 0,
-            message: event.data.error
-          })
-          reject(new Error(event.data.error))
-          break
-      }
-    })
-    
-    worker.addEventListener('error', (error) => {
-      worker.terminate()
-      onProgress?.({
-        status: 'error',
-        progress: 0,
-        message: error.message
-      })
-      reject(error)
-    })
-    
-    // Start compression
-    worker.postMessage({
-      id,
-      action: 'compress',
-      data: { content }
-    })
-  })
-}
