--- test-compression.html
@@ -1,152 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Compression Test</title>
-    <script src="https://unpkg.com/pako@2.1.0/dist/pako.min.js"></script>
-    <style>
-        body { font-family: monospace; padding: 20px; max-width: 800px; margin: 0 auto; }
-        .test { margin: 20px 0; padding: 10px; border: 1px solid #ccc; }
-        .success { background: #e8f5e9; }
-        .info { background: #e3f2fd; }
-        textarea { width: 100%; height: 100px; }
-    </style>
-</head>
-<body>
-    <h1>Markdown Compression Test</h1>
-    
-    <div class="test">
-        <h2>Test Content</h2>
-        <textarea id="testContent"># Test Document
-
-This is a test markdown document with various content types.
-
-## Code Example
-
-```javascript
-function hello() {
-    console.log("Hello, world!");
-}
-```
-
-## List
-
-- Item 1
-- Item 2
-- Item 3
-
-## Table
-
-| Column 1 | Column 2 |
-|----------|----------|
-| Data 1   | Data 2   |</textarea>
-        <button onclick="runTests()">Run Compression Tests</button>
-    </div>
-    
-    <div id="results"></div>
-    
-    <script>
-        // Check for native compression support
-        const hasNativeCompression = (() => {
-            try {
-                new CompressionStream('gzip');
-                return true;
-            } catch {
-                return false;
-            }
-        })();
-        
-        // Base64 URL encoding
-        function toBase64URL(uint8Array) {
-            let binary = '';
-            uint8Array.forEach(byte => binary += String.fromCharCode(byte));
-            let base64 = btoa(binary);
-            return base64.replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
-        }
-        
-        // Native compression
-        async function compressNative(text) {
-            const cs = new CompressionStream('gzip');
-            const byteData = new TextEncoder().encode(text);
-            const stream = new Response(byteData).body;
-            const compressedStream = stream.pipeThrough(cs);
-            const compressedBlob = await new Response(compressedStream).blob();
-            return new Uint8Array(await compressedBlob.arrayBuffer());
-        }
-        
-        // Pako compression
-        function compressPako(text) {
-            return pako.gzip(text, { level: 6 });
-        }
-        
-        async function runTests() {
-            const content = document.getElementById('testContent').value;
-            const results = document.getElementById('results');
-            results.innerHTML = '';
-            
-            // Original size
-            const originalSize = new Blob([content]).size;
-            addResult('Original Size', `${originalSize} bytes`, 'info');
-            
-            // Test native compression
-            if (hasNativeCompression) {
-                const start = performance.now();
-                const compressed = await compressNative(content);
-                const encoded = toBase64URL(compressed);
-                const time = performance.now() - start;
-                
-                addResult('Native Compression', 
-                    `Compressed: ${compressed.length} bytes, ` +
-                    `Encoded: ${encoded.length} chars, ` +
-                    `Ratio: ${Math.round((1 - compressed.length / originalSize) * 100)}%, ` +
-                    `Time: ${time.toFixed(2)}ms`, 
-                    'success'
-                );
-            } else {
-                addResult('Native Compression', 'Not supported in this browser', 'info');
-            }
-            
-            // Test Pako compression
-            {
-                const start = performance.now();
-                const compressed = compressPako(content);
-                const encoded = toBase64URL(compressed);
-                const time = performance.now() - start;
-                
-                addResult('Pako Compression', 
-                    `Compressed: ${compressed.length} bytes, ` +
-                    `Encoded: ${encoded.length} chars, ` +
-                    `Ratio: ${Math.round((1 - compressed.length / originalSize) * 100)}%, ` +
-                    `Time: ${time.toFixed(2)}ms`, 
-                    'success'
-                );
-            }
-            
-            // URL length test
-            const baseURL = window.location.origin + '/#share/';
-            const maxURL = 32000;
-            const pakoEncoded = toBase64URL(compressPako(content));
-            const totalLength = baseURL.length + pakoEncoded.length;
-            
-            addResult('URL Length Test', 
-                `Base URL: ${baseURL.length} chars, ` +
-                `Encoded content: ${pakoEncoded.length} chars, ` +
-                `Total: ${totalLength} chars, ` +
-                `Within limit (${maxURL}): ${totalLength < maxURL ? 'Yes ✓' : 'No ✗'}`,
-                totalLength < maxURL ? 'success' : 'error'
-            );
-        }
-        
-        function addResult(title, text, className) {
-            const div = document.createElement('div');
-            div.className = 'test ' + className;
-            div.innerHTML = `<strong>${title}:</strong> ${text}`;
-            document.getElementById('results').appendChild(div);
-        }
-        
-        // Run tests on load
-        window.onload = () => runTests();
-    </script>
-</body>
-</html>
